requirements.txt
bittensor>=6.0.0
substrate-interface>=1.7.0
websocket-client>=1.6.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
numpy>=1.24.0
asyncio-mqtt>=0.16.0
websockets>=12.0
aiohttp>=3.9.0
